BUGFIX: expire bfqq if needed in bfq_bfqq_move
If bfqq happens to be in service but empty when bfq_bfqq_move is invoked,
then bfqq is also still marked as busy and bfq_del_bfqq_busy is invoked.
This triggers the failure of the check that bfq_del_bfqq_busy is not
invoked on the in-service queue. In particular, this flow of operations
is imprecise also because it is better to expire and reschedule bfqq
if in service, so as to make sure that it continues to be served only
if it makes sense accordi to its new group memebrship.

This commit fixes these problems by just expring bfqq, as a preliminary
step, if bfqq is in service.

(cherry picked from commit 67e91b77c5f455e8ce3fe8728e0d4ba877e2e302)
(cherry picked from commit f9734ca6682894fe31104905b5af25147e2e4398)